security: Moved cache dropping function
